Cart-Users-featuress

+++ src/main/java/org/dani/ecommerce/services/ProductService.java
@@ -1,7 +1,10 @@
 package org.dani.ecommerce.services;
 
+import org.dani.ecommerce.enums.Role;
 import org.dani.ecommerce.models.ProductModel;
+import org.dani.ecommerce.models.UserModel;
 import org.dani.ecommerce.repositories.ProductRepo;
+import org.dani.ecommerce.repositories.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
 import java.util.List;
@@ -11,12 +14,25 @@ import java.util.List;
 public class ProductService {
 
     private ProductRepo productRepo;
+    private final UserRepository userRepository;
 
-    public ProductService(ProductRepo productRepo) {
+    public ProductService(UserRepository userRepository, ProductRepo productRepo) {
+        this.userRepository = userRepository;
         this.productRepo = productRepo;
     }
 
-    public ProductModel createProduct(ProductModel productModel){
+
+    public ProductModel createProduct(ProductModel productModel, String sellerUuid) {
+
+        UserModel user = userRepository.findByUuid(sellerUuid)
+                .orElseThrow(() -> new RuntimeException("User not found"));
+
+        if (user.getRole() != Role.SELLER) {
+            throw new RuntimeException("Only SELLER can create products");
+        }
+
+
+        productModel.setSellerUuid(user.getUuid());
         return productRepo.save(productModel);
     }
 
@@ -24,14 +40,19 @@ public class ProductService {
         return productRepo.findAll();
     }
 
-    public ProductModel updateProduct (Long id, ProductModel productModel){
+    public ProductModel updateProduct (Long id, ProductModel productModel, String sellerUuid) {
         ProductModel existProduct = productRepo.findById(id)
                 .orElseThrow(()->new RuntimeException("product not found"));
+
+        if (!existProduct.getSellerUuid().equals(sellerUuid)) {
+            throw new RuntimeException("You are not allowed to update this product");
+        }
 
         existProduct.setProductName(productModel.getProductName());
         existProduct.setProductCategory(productModel.getProductCategory());
         existProduct.setProductPrice(productModel.getProductPrice());
         existProduct.setProductDescription(productModel.getProductDescription());
+        existProduct.setProductImageUrl(productModel.getProductImageUrl());
 
         return productRepo.save(existProduct);
     }
@@ -41,9 +62,14 @@ public class ProductService {
                 .orElseThrow(()->new RuntimeException("product not found" + id));
     }
 
-    public ProductModel deleteProduct(Long id){
+    public ProductModel deleteProduct(Long id, String sellerUuid){
+
         ProductModel product =  productRepo.findById(id)
                 .orElseThrow(()-> new RuntimeException("product not found" + id));
+
+        if (!product.getSellerUuid().equals(sellerUuid)) {
+            throw new RuntimeException("You are not allowed to delete this product");
+        }
 
         productRepo.delete(product);
         return product;
